+++ src/layer.rs
@@ -7,8 +7,7 @@ use tracing::{span, Subscriber};
 use tracing_subscriber::{registry::LookupSpan, Layer};
 
 use crate::activities::Activities; // filter::{FilterFn}
-use crate::native::*;
-use crate::providerwrapper::{map_level, GuidWrapper, ProviderGroup};
+use crate::{native::*, map_level};
 use crate::values::*;
 
 // Providers go in, but never come out.
